+++ native-windows-wysiwyg/src/gui.rs
@@ -1,3 +1,4 @@
+mod gui_shared;
 mod gui_error;
 mod widget_box;
 mod object_inspector;