+++ native-windows-wysiwyg/src/gui/object_inspector.rs
@@ -1,5 +1,6 @@
 use nwd::NwgPartial;
-use std::cell::Cell;
+use std::{cell::Cell, rc::Rc};
+use super::gui_shared::set_cursor;
 
 
 #[derive(Default)]
@@ -9,10 +10,20 @@ pub struct ObjectInspector {
     pub(super) user_width: Cell<u32>,
 
     // Saved control list height
-    control_list_height: Cell<u32>,
+    control_list_height: Rc<Cell<u32>>,
+
+    #[nwg_resource(source_system: Some(nwg::OemCursor::SizeWE))]
+    cursor_size_we: nwg::Cursor,
+
+    #[nwg_resource(source_system: Some(nwg::OemCursor::Normal))]
+    cursor_default: nwg::Cursor,
 
     #[nwg_control(size: (275, 0))]
-    #[nwg_events(OnResize: [ObjectInspector::resize])]
+    #[nwg_events(
+        OnResize: [ObjectInspector::resize],
+        OnMouseMove: [ObjectInspector::update_cursor],
+        OnMouseLeft: [ObjectInspector::clear_cursor],
+    )]
     pub(super) container_frame: nwg::Frame,
 
     #[nwg_control(parent: container_frame, text: "Controls", v_align: nwg::VTextAlign::Top)]
@@ -38,6 +49,7 @@ pub struct ObjectInspector {
 
 impl ObjectInspector {
 
+
     pub(super) fn init(&self) {
         self.user_width.set(275);
         self.control_list_height.set(300);
@@ -59,6 +71,7 @@ impl ObjectInspector {
         let (w, h) = self.container_frame.size();
         let left_offset = 5i32;
         let label_height = 25;
+        let control_list_height = self.control_list_height.get();
         let width_with_offset = w - ((left_offset as u32) * 2);
 
         let mut top_offset = 5;
@@ -68,8 +81,8 @@ impl ObjectInspector {
         top_offset += label_height as i32;
 
         self.control_list.set_position(left_offset, top_offset);
-        self.control_list.set_size(width_with_offset, self.control_list_height.get());
-        top_offset += (5 + self.control_list_height.get()) as i32;
+        self.control_list.set_size(width_with_offset, control_list_height);
+        top_offset += (5 + control_list_height) as i32;
 
         self.properties_label.set_position(left_offset, top_offset);
         self.properties_label.set_size(width_with_offset, label_height);
@@ -77,6 +90,35 @@ impl ObjectInspector {
 
         self.properties_list.set_position(left_offset, top_offset);
         self.properties_list.set_size(width_with_offset, h - (top_offset as u32) - 10);
+    }
+
+    fn update_cursor(&self) {
+        let (x, _y) = nwg::GlobalCursor::local_position(&self.container_frame, None);
+        
+        /*
+        // Safe "recommended" (by the windows docs) way to set the cursor. Flickers a bit.
+        if x < 4 {
+            nwg::GlobalCursor::set(&self.size_we);
+        }
+        */
+
+        // This tells winapi to send a `Event::OnMouseLeave` when the cursor will leave the control
+        nwg::GlobalCursor::track_mouse_leaving(&self.container_frame);
+
+        // The ~cool~ way to set a cursor in winapi that never flickers
+        // Sadly (for me, the dev of NWG) it's impossible to safely wrap because this
+        // sets the cursor ON ALL INSTANCES OF THE SAME CONTROL. Goddamit Windows.
+        if x < 5 {
+            set_cursor(&self.container_frame.handle, &self.cursor_size_we);
+        } else {
+            set_cursor(&self.container_frame.handle, &self.cursor_default);
+        }
+    }
+
+    /// This gets triggered by the `track_mouse_leaving` call in `update_cursor`
+    /// Sets the cursor back to normal so that the other frame works correctly
+    fn clear_cursor(&self) {
+        set_cursor(&self.container_frame.handle, &self.cursor_default);
     }
 
 }