+++ native-windows-wysiwyg/src/gui/widget_box.rs
@@ -1,5 +1,6 @@
 use nwd::NwgPartial;
 use std::cell::Cell;
+use super::gui_shared::set_cursor;
 
 #[derive(Default)]
 #[derive(NwgPartial)]
@@ -7,10 +8,20 @@ pub struct WidgetBox {
     // Saved width to keep then restore when the main window is too small
     pub(super) user_width: Cell<u32>,
 
+    #[nwg_resource(source_system: Some(nwg::OemCursor::SizeWE))]
+    cursor_size_we: nwg::Cursor,
+
+    #[nwg_resource(source_system: Some(nwg::OemCursor::Normal))]
+    cursor_default: nwg::Cursor,
+
     #[nwg_control(size: (275, 0))]
+    #[nwg_events(
+        OnMouseMove: [WidgetBox::update_cursor],
+        OnMouseLeft: [WidgetBox::clear_cursor],
+    )]
     pub(super) container_frame: nwg::Frame,
 
-    #[nwg_layout(parent: container_frame, spacing: 0, margin: [0,0,0,0])]
+    #[nwg_layout(parent: container_frame, spacing: 0, margin: [0,5,0,0])]
     layout: nwg::GridLayout,
 
     #[nwg_control(parent: container_frame)]
@@ -22,6 +33,7 @@ impl WidgetBox {
 
     pub(super) fn init(&self) {
         self.user_width.set(275);
+        self.load_widgets();
     }
 
     pub(super) fn load_widgets(&self) {
@@ -74,7 +86,7 @@ impl WidgetBox {
         let other = tree.insert_item("Other", Some(&controls), nwg::TreeInsert::Last);
         tree.insert_item("Track bar", Some(&other), nwg::TreeInsert::Last);
 
-        let _custom = tree.insert_item("Custom", Some(&controls), nwg::TreeInsert::Last);
+        let _custom = tree.insert_item("Partials", Some(&controls), nwg::TreeInsert::Last);
 
         for item in tree.iter() {
             tree.set_expand_state(&item, nwg::ExpandState::Expand);
@@ -84,4 +96,31 @@ impl WidgetBox {
         tree.set_enabled(false);
     }
 
+
+    fn update_cursor(&self) {
+        let (x, _y) = nwg::GlobalCursor::local_position(&self.container_frame, None);
+        let (w, _h) = self.container_frame.size();
+        
+        /*
+        // Safe "recommended" (by the windows docs) way to set the cursor. Flickers a bit.
+        if x < 4 {
+            nwg::GlobalCursor::set(&self.size_we);
+        }
+        */
+
+        // This tells winapi to send a `Event::OnMouseLeave` when the cursor will leave the control
+        nwg::GlobalCursor::track_mouse_leaving(&self.container_frame);
+
+        if x > (w as i32 - 5) {
+            set_cursor(&self.container_frame.handle, &self.cursor_size_we);
+        } else {
+            set_cursor(&self.container_frame.handle, &self.cursor_default);
+        }
+    }
+
+    /// This gets triggered by the `track_mouse_leaving` call in `update_cursor`
+    /// Sets the cursor back to normal so that the other frame works correctly
+    fn clear_cursor(&self) {
+        set_cursor(&self.container_frame.handle, &self.cursor_default);
+    }
 }