+++ changelog.md
@@ -1,7 +1,11 @@
+1.0.11
+* Fixed `Frame` control `set_size`
+* Added `track_mouse_leaving`, `track_mouse_hover`, `track_mouse_cancel`, and `track_mouse_query` to `GlobalCursor`
+
 1.0.10
 * Fixed a compiling bug when using `no-default-features`
 
-1.0.9
+1.0.9 (yanked, use 1.0.10)
 * BREAKING CHANGE: File dialog `get_selected_item` & `get_selected_items` now return `OsString` instead of String to handle some exotic Windows path
 * Added vertical alignment  `v_align` to label (defaults to center)
 * Added vertical alignment `v_align` to combobox (defaults to center)